Fix format warnings in libr/core

+++ libr/core/cmd_api.c
@@ -520,7 +520,7 @@ static size_t update_max_len(RCmdDesc *cd, size_t max_len) {
 
 static void print_child_help(RStrBuf *sb, RCmdDesc *cd, size_t max_len, bool use_color) {
 	size_t str_len = calc_padding_len (cd);
-	size_t padding = str_len < max_len? max_len - str_len: 0;
+	int padding = str_len < max_len? max_len - str_len: 0;
 	const char *cd_summary = cd->help->summary? cd->help->summary: "";
 
 	RCons *cons = r_cons_singleton ();
